refactor routes and handlers, health check endpoint
