+++ .env
@@ -1,5 +1,5 @@
 BRIZO_DB_USER=root
 BRIZO_DB_PASSWORD=secret
-BRIZO_DB_HOST=127.0.0.1
+BRIZO_DB_HOST=localhost
 BRIZO_DB_PORT=33062
 BRIZO_DB_NAME=brizo
