+++ database/connection_test.go
@@ -14,7 +14,7 @@ func TestGetDBAddress(t *testing.T) {
 	os.Setenv("BRIZO_DB_PORT", "3307")
 	os.Setenv("BRIZO_DB_NAME", "somedb")
 
-	expected := "foo:bar@db.example.com:3307/somedb?charset=utf8&parseTime=True&loc=Local"
+	expected := "foo:bar@(db.example.com:3307)/somedb?charset=utf8&parseTime=True&loc=Local"
 	actual := getDBAddress()
 	assert.Equal(t, expected, actual)
 }
@@ -26,7 +26,7 @@ func TestGetDBAddressUsesDefaultPort(t *testing.T) {
 	os.Setenv("BRIZO_DB_PORT", "")
 	os.Setenv("BRIZO_DB_NAME", "somedb")
 
-	expected := "foo:bar@db.example.com:3306/somedb?charset=utf8&parseTime=True&loc=Local"
+	expected := "foo:bar@(db.example.com:3306)/somedb?charset=utf8&parseTime=True&loc=Local"
 	actual := getDBAddress()
 	assert.Equal(t, expected, actual)
 }
