+++ docker-compose.yml
@@ -3,7 +3,7 @@ services:
   database:
     image: mysql:5.6
     ports:
-      - 33061:3306
+      - 33062:3306
     environment:
       MYSQL_ROOT_PASSWORD: secret
       MYSQL_DATABASE: brizo
