+++ runner.conf
@@ -2,7 +2,7 @@ root:              .
 tmp_path:          ./tmp
 build_name:        runner-build
 build_log:         runner-build-errors.log
-valid_ext:         .go, .jet
+valid_ext:         .go, .jet, .env
 ignored:           ui
 build_delay:       600
 colors:            1
